Move more functions to posix.cs
